Moved models into separate namespace
