+++ src/LinkyLink/GetLinks.cs
@@ -6,6 +6,7 @@ using Microsoft.Extensions.Logging;
 using System.Collections.Generic;
 using System.Linq;
 using Microsoft.Azure.Documents;
+using LinkyLink.Models;
 
 namespace LinkyLink
 {
